pushRemoteStatus: Skip a bunch of network calls
This ended up grinding things to total halt, when the remote and local
was far off each other (though I didn't see this before today, and had
used the same branch earlier, though perhaps before I rebased?).

So the vscode git api is not great, it breaks things up too much making
it hard to do things efficently (see the diff stuff, where
you're not able to get the whole diff from eg. a commit in one call).

So first off, we probably shouldn't try to fetch at all on status,
it's in the ui fast path and we don't want to wait for the network.

Second, running `getCommit` on each sha looks to be very slow, I'm
guessing it ends translating to a git process for each sha and what
looks like network traffic.